navigate to input pwd screen + extract game tile
