+++ lib/screens/lobby/created_game_tile.dart
@@ -1,75 +1,20 @@
 import 'package:flutter/material.dart';
 import 'package:niira/models/game.dart';
 import 'package:niira/navigation/navigation.dart';
-import 'package:niira/services/auth/auth_service.dart';
-import 'package:niira/services/database/database_service.dart';
 import 'package:niira/services/game_service.dart';
 import 'package:provider/provider.dart';
 
-class LobbyScreen extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text('Lobby'),
-        actions: [
-          FlatButton(
-            key: Key('signOutBtn'),
-            onPressed: () {
-              // create a function to call on confirmation
-              final signOut = () async {
-                context.read<Navigation>().pop();
-                await context.read<AuthService>().signOut();
-              };
-
-              context.read<Navigation>().showConfirmationDialog(
-                    onConfirmed: signOut,
-                    confirmText: 'Sign Out',
-                    cancelText: 'Return',
-                  );
-            },
-            child: Text('log out'),
-          )
-        ],
-      ),
-      body: Container(
-        child: StreamBuilder<List<Game>>(
-            stream: context.watch<DatabaseService>().streamOfCreatedGames,
-            builder: (context, snapshot) {
-              if (snapshot.data == null) {
-                return Container();
-              } else {
-                return Padding(
-                  padding: const EdgeInsets.fromLTRB(0, 10, 0, 0),
-                  child: ListView.builder(
-                    itemCount: snapshot.data.length,
-                    itemBuilder: (context, index) {
-                      return GameTile(
-                        snapshot.data[index],
-                      );
-                    },
-                  ),
-                );
-              }
-            }),
-      ),
-      floatingActionButton: FloatingActionButton.extended(
-        onPressed: () => context.read<Navigation>().navigateTo('/new_game1'),
-        label: Text('New Game'),
-        icon: Icon(Icons.add),
-      ),
-    );
-  }
-}
-
 class GameTile extends StatelessWidget {
   final Game _game;
-  GameTile(this._game);
+  final int _index;
+  GameTile(this._game, this._index);
 
   @override
   Widget build(BuildContext context) {
     return Container(
-      key: Key('created_game_tile_${_game.id}'),
+      // key is formatted with index to help with testing,
+      // refer to created_game_tile_widget_test.dart
+      key: Key('created_game_tile_${_game.id}_index:${_index}'),
       child: Padding(
         padding: EdgeInsets.fromLTRB(20, 5, 20, 5),
         child: Card(
@@ -86,18 +31,21 @@ class GameTile extends StatelessWidget {
                     crossAxisAlignment: CrossAxisAlignment.start,
                     children: <Widget>[
                       Text(
-                        '500m away', //TODO: will replace with real data in ticket #60
+                        '500m away',
                         textAlign: TextAlign.left,
                         style: TextStyle(
                           fontSize: 14,
                           color: Colors.black.withOpacity(0.6),
                         ),
                       ),
-                      Text(
-                        _game.name, // title of game
-                        textAlign: TextAlign.left,
-                        style: TextStyle(
-                          fontSize: 24,
+                      Padding(
+                        padding: EdgeInsets.fromLTRB(0, 2.5, 0, 2.5),
+                        child: Text(
+                          _game.name, // title of game
+                          textAlign: TextAlign.left,
+                          style: TextStyle(
+                            fontSize: 24,
+                          ),
                         ),
                       ),
                       Text(
