+++ test/create_game_section/Lobby/lobby_widget_test.dart
@@ -9,7 +9,7 @@ import 'package:niira/models/game.dart';
 import 'package:niira/models/user_data.dart';
 import 'package:niira/navigation/navigation.dart';
 import 'package:niira/screens/input_password.dart';
-import 'package:niira/screens/lobby.dart';
+import 'package:niira/screens/lobby/lobby.dart';
 import 'package:niira/services/database/database_service.dart';
 import 'package:niira/services/game_service.dart';
 import 'package:provider/provider.dart';
